td-shim: refine memory related code
Refine memory related code by spliting TDX specific code into a
dedicated submodule.
