+++ rust-tdshim/src/td/mod.rs
@@ -1,0 +1,7 @@
+// Copyright (c) 2022 Alibaba Cloud
+//
+// SPDX-License-Identifier: BSD-2-Clause-Patent
+
+mod tdx;
+
+pub use tdx::*;
